+++ src/module-registrar.cc
@@ -661,16 +661,13 @@ bool ModuleRegistrar::isManagedDomain(const url_t* url) {
 
 void ModuleRegistrar::removeInternalParams(sip_contact_t* ct) {
 	for (sip_contact_t* contact = ct; contact != nullptr; contact = contact->m_next) {
-		if (contact) {
-			if (url_has_param(contact->m_url, "fs-conn-id")) {
-				contact->m_url->url_params = url_strip_param_string((char*)contact->m_url->url_params, "fs-conn-id");
-			}
-			const char* pub_gruu_value = msg_header_find_param((msg_common_t*)contact, "pub-gruu");
-			if (pub_gruu_value && pub_gruu_value[0] == '\0') {
-				/* Remove empty pub-gruu parameter, which is set internally
-				 * for compatibility with previous gruu implementation.*/
-				msg_header_remove_param((msg_common_t*)contact, "pub-gruu");
-			}
+		msg_header_remove_param(reinterpret_cast<msg_common_t*>(ct), "fs-conn-id");
+
+		const auto* pubGruu = msg_header_find_param(reinterpret_cast<msg_common_t*>(contact), "pub-gruu");
+		if (pubGruu && pubGruu[0] == '\0') {
+			// Remove the empty 'pub-gruu' parameter, which is set internally for compatibility with the previous gruu
+			// implementation.
+			msg_header_remove_param(reinterpret_cast<msg_common_t*>(contact), "pub-gruu");
 		}
 	}
 }
@@ -855,12 +852,13 @@ unique_ptr<RequestSipEvent> ModuleRegistrar::onRequest(unique_ptr<RequestSipEven
 	/* Initialize a connection ID, so that registration can be matched with the tport,
 	   in order to later identify aborted connections during subsequent registrations. */
 	{
-		ostringstream os;
 		uintptr_t connId = (tport_get_user_data(ev->getIncomingTport().get()))
 		                       ? reinterpret_cast<uintptr_t>(tport_get_user_data(ev->getIncomingTport().get()))
 		                       : static_cast<uintptr_t>(su_random64());
-		os << "fs-conn-id=" << hex << connId;
-		url_param_add(ms->getHome(), sip->sip_contact->m_url, os.str().c_str());
+		ostringstream param{};
+		param << "fs-conn-id=" << hex << connId;
+		msg_header_add_param(ms->getHome(), reinterpret_cast<msg_common_t*>(sip->sip_contact),
+		                     su_strdup(ms->getHome(), param.str().c_str()));
 		tport_set_user_data(ev->getIncomingTport().get(), reinterpret_cast<void*>(connId));
 	}